cleaned up code structure
