+++ src/components/stickies.js
@@ -1,7 +1,19 @@
 import React, { useState } from "react";
 
-export default function Sticky(props) {
+export default function Sticky() {
   const [text, handleTextChange] = useState("");
+  const [isVisible, setVisibility] = useState("flex");
+
+  const toggleVisibility = () => {
+    if (
+      window.confirm(
+        "Are you sure you want to delete this note? This action cannot be undone."
+      )
+    ) {
+      setVisibility("none");
+      localStorage.removeItem("inputValue");
+    }
+  };
 
   const handleChange = (e) => {
     handleTextChange(e.target.value);
@@ -10,10 +22,10 @@ export default function Sticky(props) {
 
   return (
     <div>
-      <ul style={props.style} className="sticky">
+      <ul style={{ display: isVisible }} className="sticky">
         <li>
           <div className="button-container">
-            <button onClick={props.onClick} id="close-button">
+            <button onClick={() => toggleVisibility()} id="close-button">
               x
             </button>
           </div>
